No more NULL dereference and kernel panic on current kernels
Found patch on the web:
  http://users.telenet.be/x86_64/Patches/rtl8192eu-k3.13.patch

Before patch, kernel panic due to a NULL dereference in
rtw_ndev_notifier_register; this patch solves that problem by not using
these methods on kernel versions after a certain number, so it leads me
to believe that this patch is meant to make the driver compatible with
newer kernel versions.

+++ os_dep/linux/rtw_proc.c
@@ -319,7 +319,7 @@ int proc_get_rx_info(struct seq_file *m, void *v)
 }	
 
 
-int proc_reset_rx_info(struct file *file, const char __user *buffer, size_t count, loff_t *pos, void *data)
+ssize_t proc_reset_rx_info(struct file *file, const char __user *buffer, size_t count, loff_t *pos, void *data)
 {
 	struct net_device *dev = data;
 	_adapter *padapter = (_adapter *)rtw_netdev_priv(dev);
